+++ fsm/repos/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/repos/common/KeyValueBasedStatesManager.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
-@Deprecated
+@Deprecated("Replace with DefaultStatesManager and KeyValueBasedDefaultStatesManagerRepo")
 class KeyValueBasedStatesManager(
     private val keyValueRepo: KeyValueRepo<Any, State>,
     private val onContextsConflictResolver: suspend (old: State, new: State, currentNew: State) -> Boolean = { _, _, _ -> true }